+++ src/main/java/com/owlmaddie/chat/ChatMessage.java
@@ -6,9 +6,11 @@ package com.owlmaddie.chat;
 public class ChatMessage {
     public String message;
     public ChatDataManager.ChatSender sender;
+    public Long timestamp;
 
     public ChatMessage(String message, ChatDataManager.ChatSender sender) {
         this.message = message;
         this.sender = sender;
+        this.timestamp = System.currentTimeMillis();
     }
 }